LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/return-value/traverseTo-repeated-expected.txt


FAIL Repeated navigation.traverseTo() with the same key assert_equals: committed promises must be equal expected object "[object Promise]" but got object "[object Promise]"
